pyproject.toml
[tool.poetry]
name = "openai-realtime-client"
version = "0.1.0"
description = "A python-based client for OpenAI's Realtime API"
authors = ["Logan Markewich <logan@runllama.ai>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
llama-index-core = "^0.11.17"
pyaudio = "^0.2.14"
pynput = "^1.7.7"
pydub = "^0.25.1"
websockets = "^13.1"
wave = "^0.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
